+++ src/gibbs_functions.cpp
@@ -1,5 +1,6 @@
 // [[Rcpp::depends(RcppProgress)]]
 #include <Rcpp.h>
+#include "gibbs_functions_edge_prior.h"
 #include <progress.hpp>
 #include <progress_bar.hpp>
 using namespace Rcpp;
@@ -111,15 +112,14 @@ List impute_missing_data(NumericMatrix interactions,
                       Named("n_cat_obs") = n_cat_obs,
                       Named("sufficient_blume_capel") =
                         sufficient_blume_capel,
-                      Named("rest_matrix") = rest_matrix);
+                        Named("rest_matrix") = rest_matrix);
 }
 
 // ----------------------------------------------------------------------------|
 // MH algorithm to sample from the full-conditional of the threshold parameters
 //   for a regular binary or ordinal variable
 // ----------------------------------------------------------------------------|
-void metropolis_thresholds_regular(NumericMatrix interactions,
-                                   NumericMatrix thresholds,
+void metropolis_thresholds_regular(NumericMatrix thresholds,
                                    IntegerMatrix observations,
                                    IntegerVector no_categories,
                                    IntegerMatrix n_cat_obs,
@@ -193,8 +193,7 @@ void metropolis_thresholds_regular(NumericMatrix interactions,
 // Adaptive Metropolis algorithm to sample from the full-conditional of the
 //   threshold parameters for a Blume-Capel ordinal variable
 // ----------------------------------------------------------------------------|
-void metropolis_thresholds_blumecapel(NumericMatrix interactions,
-                                      NumericMatrix thresholds,
+void metropolis_thresholds_blumecapel(NumericMatrix thresholds,
                                       IntegerMatrix observations,
                                       IntegerVector no_categories,
                                       IntegerMatrix sufficient_blume_capel,
@@ -496,7 +495,7 @@ double log_pseudolikelihood_ratio(NumericMatrix interactions,
 // ----------------------------------------------------------------------------|
 void metropolis_interactions(NumericMatrix interactions,
                              NumericMatrix thresholds,
-                             IntegerMatrix gamma,
+                             IntegerMatrix indicator,
                              IntegerMatrix observations,
                              IntegerVector no_categories,
                              NumericMatrix proposal_sd,
@@ -518,7 +517,7 @@ void metropolis_interactions(NumericMatrix interactions,
 
   for(int variable1 = 0; variable1 <  no_variables - 1; variable1++) {
     for(int variable2 = variable1 + 1; variable2 <  no_variables; variable2++) {
-      if(gamma(variable1, variable2) == 1) {
+      if(indicator(variable1, variable2) == 1) {
         current_state = interactions(variable1, variable2);
         proposed_state = R::rnorm(current_state, proposal_sd(variable1, variable2));
 
@@ -575,7 +574,7 @@ void metropolis_interactions(NumericMatrix interactions,
 // ----------------------------------------------------------------------------|
 void metropolis_edge_interaction_pair(NumericMatrix interactions,
                                       NumericMatrix thresholds,
-                                      IntegerMatrix gamma,
+                                      IntegerMatrix indicator,
                                       IntegerMatrix observations,
                                       IntegerVector no_categories,
                                       NumericMatrix proposal_sd,
@@ -601,7 +600,7 @@ void metropolis_edge_interaction_pair(NumericMatrix interactions,
 
     current_state = interactions(variable1, variable2);
 
-    if(gamma(variable1, variable2) == 0) {
+    if(indicator(variable1, variable2) == 0) {
       proposed_state = R::rnorm(current_state, proposal_sd(variable1, variable2));
     } else {
       proposed_state = 0.0;
@@ -620,7 +619,7 @@ void metropolis_edge_interaction_pair(NumericMatrix interactions,
                                           variable_bool,
                                           reference_category);
 
-    if(gamma(variable1, variable2) == 0) {
+    if(indicator(variable1, variable2) == 0) {
       log_prob += R::dcauchy(proposed_state, 0.0, interaction_scale, true);
       log_prob -= R::dnorm(proposed_state,
                            current_state,
@@ -640,8 +639,8 @@ void metropolis_edge_interaction_pair(NumericMatrix interactions,
 
     U = R::unif_rand();
     if(std::log(U) < log_prob) {
-      gamma(variable1, variable2) = 1 - gamma(variable1, variable2);
-      gamma(variable2, variable1) = 1 - gamma(variable2, variable1);
+      indicator(variable1, variable2) = 1 - indicator(variable1, variable2);
+      indicator(variable2, variable1) = 1 - indicator(variable2, variable1);
 
       interactions(variable1, variable2) = proposed_state;
       interactions(variable2, variable1) = proposed_state;
@@ -677,7 +676,7 @@ List gibbs_step_gm(IntegerMatrix observations,
                    int no_interactions,
                    int no_thresholds,
                    int max_no_categories,
-                   IntegerMatrix gamma,
+                   IntegerMatrix indicator,
                    NumericMatrix interactions,
                    NumericMatrix thresholds,
                    NumericMatrix rest_matrix,
@@ -695,7 +694,7 @@ List gibbs_step_gm(IntegerMatrix observations,
     //Between model move (update edge indicators and interaction parameters)
     metropolis_edge_interaction_pair(interactions,
                                      thresholds,
-                                     gamma,
+                                     indicator,
                                      observations,
                                      no_categories,
                                      proposal_sd,
@@ -712,7 +711,7 @@ List gibbs_step_gm(IntegerMatrix observations,
   //Within model move (update interaction parameters)
   metropolis_interactions(interactions,
                           thresholds,
-                          gamma,
+                          indicator,
                           observations,
                           no_categories,
                           proposal_sd,
@@ -731,8 +730,7 @@ List gibbs_step_gm(IntegerMatrix observations,
   //Update threshold parameters
   for(int variable = 0; variable < no_variables; variable++) {
     if(variable_bool[variable] == true) {
-      metropolis_thresholds_regular(interactions,
-                                    thresholds,
+      metropolis_thresholds_regular(thresholds,
                                     observations,
                                     no_categories,
                                     n_cat_obs,
@@ -742,8 +740,7 @@ List gibbs_step_gm(IntegerMatrix observations,
                                     threshold_beta,
                                     rest_matrix);
     } else {
-      metropolis_thresholds_blumecapel(interactions,
-                                       thresholds,
+      metropolis_thresholds_blumecapel(thresholds,
                                        observations,
                                        no_categories,
                                        sufficient_blume_capel,
@@ -762,7 +759,7 @@ List gibbs_step_gm(IntegerMatrix observations,
     }
   }
 
-  return List::create(Named("gamma") = gamma,
+  return List::create(Named("indicator") = indicator,
                       Named("interactions") = interactions,
                       Named("thresholds") = thresholds,
                       Named("rest_matrix") = rest_matrix,
@@ -774,7 +771,7 @@ List gibbs_step_gm(IntegerMatrix observations,
 // ----------------------------------------------------------------------------|
 // [[Rcpp::export]]
 List gibbs_sampler(IntegerMatrix observations,
-                   IntegerMatrix gamma,
+                   IntegerMatrix indicator,
                    NumericMatrix interactions,
                    NumericMatrix thresholds,
                    IntegerVector no_categories,
@@ -785,6 +782,7 @@ List gibbs_sampler(IntegerMatrix observations,
                    NumericMatrix theta,
                    double beta_bernoulli_alpha,
                    double beta_bernoulli_beta,
+                   double dirichlet_alpha,
                    IntegerMatrix Index,
                    int iter,
                    int burnin,
@@ -814,7 +812,7 @@ List gibbs_sampler(IntegerMatrix observations,
   double phi = .75;
   double target_ar = 0.234;
   double epsilon_lo = 1 / no_persons;
-  double epsilon_hi = 2.0;
+  double epsilon_hi = 20.0;
 
   //The resizing based on ``save'' could probably be prettier ------------------
   int nrow = no_variables;
@@ -833,14 +831,14 @@ List gibbs_sampler(IntegerMatrix observations,
   if(edge_selection == false) {
     for(int variable1 = 0; variable1 < no_variables - 1; variable1++) {
       for(int variable2 = variable1 + 1; variable2 < no_variables; variable2++) {
-        gamma(variable1, variable2) = 1;
-        gamma(variable2, variable1) = 1;
+        indicator(variable1, variable2) = 1;
+        indicator(variable2, variable1) = 1;
       }
     }
     nrow = 1;
     ncol_edges = 1;
   }
-  NumericMatrix out_gamma(nrow, ncol_edges);
+  NumericMatrix out_indicator(nrow, ncol_edges);
 
   NumericMatrix rest_matrix(no_persons, no_variables);
   for(int variable1 = 0; variable1 < no_variables; variable1++) {
@@ -852,6 +850,41 @@ List gibbs_sampler(IntegerMatrix observations,
     }
   }
 
+  //Variable declaration edge prior
+  IntegerVector cluster_allocations(no_variables);
+  NumericMatrix cluster_prob(1, 1);
+  NumericVector log_Vn(1);
+
+  if(edge_prior == "Stochastic-Block") {
+    cluster_allocations[0] = 0;
+    cluster_allocations[1] = 1;
+    for(int i = 2; i < no_variables; i++) {
+      double U = R::unif_rand();
+      if(U > 0.5){
+        cluster_allocations[i] = 1;
+      } else {
+        cluster_allocations[i] = 0;
+      }
+    }
+
+    cluster_prob = block_probs_mfm_sbm(cluster_allocations,
+                                       indicator,
+                                       no_variables,
+                                       beta_bernoulli_alpha,
+                                       beta_bernoulli_beta);
+
+    for(int i = 0; i < no_variables - 1; i++) {
+      for(int j = i + 1; j < no_variables; j++) {
+        theta(i, j) = cluster_prob(cluster_allocations[i], cluster_allocations[j]);
+        theta(j, i) = cluster_prob(cluster_allocations[i], cluster_allocations[j]);
+      }
+    }
+
+    log_Vn = compute_Vn_mfm_sbm(no_variables,
+                                dirichlet_alpha,
+                                no_variables + 10);
+  }
+
   //Progress bar
   Progress p(iter + burnin, display_progress);
 
@@ -859,7 +892,7 @@ List gibbs_sampler(IntegerMatrix observations,
   //First, we do burn-in iterations---------------------------------------------
   for(int iteration = 0; iteration < burnin; iteration++) {
     if (Progress::check_abort()) {
-      return List::create(Named("gamma") = out_gamma,
+      return List::create(Named("indicator") = out_indicator,
                           Named("interactions") = out_interactions,
                           Named("thresholds") = out_thresholds);
     }
@@ -911,7 +944,7 @@ List gibbs_sampler(IntegerMatrix observations,
                              no_interactions,
                              no_thresholds,
                              max_no_categories,
-                             gamma,
+                             indicator,
                              interactions,
                              thresholds,
                              rest_matrix,
@@ -925,7 +958,7 @@ List gibbs_sampler(IntegerMatrix observations,
                              reference_category,
                              edge_selection);
 
-    IntegerMatrix gamma = out["gamma"];
+    IntegerMatrix indicator = out["indicator"];
     NumericMatrix interactions = out["interactions"];
     NumericMatrix thresholds = out["thresholds"];
     NumericMatrix rest_matrix = out["rest_matrix"];
@@ -936,7 +969,7 @@ List gibbs_sampler(IntegerMatrix observations,
         int sumG = 0;
         for(int i = 0; i < no_variables - 1; i++) {
           for(int j = i + 1; j < no_variables; j++) {
-            sumG += gamma(i, j);
+            sumG += indicator(i, j);
           }
         }
         double probability = R::rbeta(beta_bernoulli_alpha + sumG,
@@ -949,6 +982,30 @@ List gibbs_sampler(IntegerMatrix observations,
           }
         }
       }
+      if(edge_prior == "Stochastic-Block") {
+        cluster_allocations = block_allocations_mfm_sbm(cluster_allocations,
+                                                        no_variables,
+                                                        log_Vn,
+                                                        cluster_prob,
+                                                        indicator,
+                                                        dirichlet_alpha,
+                                                        beta_bernoulli_alpha,
+                                                        beta_bernoulli_beta);
+
+
+        cluster_prob = block_probs_mfm_sbm(cluster_allocations,
+                                           indicator,
+                                           no_variables,
+                                           beta_bernoulli_alpha,
+                                           beta_bernoulli_beta);
+
+        for(int i = 0; i < no_variables - 1; i++) {
+          for(int j = i + 1; j < no_variables; j++) {
+            theta(i, j) = cluster_prob(cluster_allocations[i], cluster_allocations[j]);
+            theta(j, i) = cluster_prob(cluster_allocations[i], cluster_allocations[j]);
+          }
+        }
+      }
     }
   }
 
@@ -956,7 +1013,7 @@ List gibbs_sampler(IntegerMatrix observations,
   for(int iteration = 0; iteration < iter; iteration++) {
     if (Progress::check_abort()) {
       if(edge_selection == true) {
-        return List::create(Named("gamma") = out_gamma,
+        return List::create(Named("indicator") = out_indicator,
                             Named("interactions") = out_interactions,
                             Named("thresholds") = out_thresholds);
       } else {
@@ -1012,7 +1069,7 @@ List gibbs_sampler(IntegerMatrix observations,
                              no_interactions,
                              no_thresholds,
                              max_no_categories,
-                             gamma,
+                             indicator,
                              interactions,
                              thresholds,
                              rest_matrix,
@@ -1026,7 +1083,7 @@ List gibbs_sampler(IntegerMatrix observations,
                              reference_category,
                              edge_selection);
 
-    IntegerMatrix gamma = out["gamma"];
+    IntegerMatrix indicator = out["indicator"];
     NumericMatrix interactions = out["interactions"];
     NumericMatrix thresholds = out["thresholds"];
     NumericMatrix rest_matrix = out["rest_matrix"];
@@ -1037,7 +1094,7 @@ List gibbs_sampler(IntegerMatrix observations,
         int sumG = 0;
         for(int i = 0; i < no_variables - 1; i++) {
           for(int j = i + 1; j < no_variables; j++) {
-            sumG += gamma(i, j);
+            sumG += indicator(i, j);
           }
         }
         double probability = R::rbeta(beta_bernoulli_alpha + sumG,
@@ -1047,6 +1104,30 @@ List gibbs_sampler(IntegerMatrix observations,
           for(int j = i + 1; j < no_variables; j++) {
             theta(i, j) = probability;
             theta(j, i) = probability;
+          }
+        }
+      }
+      if(edge_prior == "Stochastic-Block") {
+        cluster_allocations = block_allocations_mfm_sbm(cluster_allocations,
+                                                        no_variables,
+                                                        log_Vn,
+                                                        cluster_prob,
+                                                        indicator,
+                                                        dirichlet_alpha,
+                                                        beta_bernoulli_alpha,
+                                                        beta_bernoulli_beta);
+
+
+        cluster_prob = block_probs_mfm_sbm(cluster_allocations,
+                                           indicator,
+                                           no_variables,
+                                           beta_bernoulli_alpha,
+                                           beta_bernoulli_beta);
+
+        for(int i = 0; i < no_variables - 1; i++) {
+          for(int j = i + 1; j < no_variables; j++) {
+            theta(i, j) = cluster_prob(cluster_allocations[i], cluster_allocations[j]);
+            theta(j, i) = cluster_prob(cluster_allocations[i], cluster_allocations[j]);
           }
         }
       }
@@ -1060,7 +1141,7 @@ List gibbs_sampler(IntegerMatrix observations,
       for(int variable1 = 0; variable1 < no_variables - 1; variable1++) {
         for(int variable2 = variable1 + 1; variable2 < no_variables; variable2++) {
           if(edge_selection == true) {
-            out_gamma(iteration, cntr) = gamma(variable1, variable2);
+            out_indicator(iteration, cntr) = indicator(variable1, variable2);
           }
           out_interactions(iteration, cntr) = interactions(variable1, variable2);
           cntr++;
@@ -1085,10 +1166,10 @@ List gibbs_sampler(IntegerMatrix observations,
       for(int variable1 = 0; variable1 < no_variables - 1; variable1++) {
         for(int variable2 = variable1 + 1; variable2 < no_variables; variable2++) {
           if(edge_selection == true) {
-            out_gamma(variable1, variable2) *= iteration;
-            out_gamma(variable1, variable2) += gamma(variable1, variable2);
-            out_gamma(variable1, variable2) /= iteration + 1;
-            out_gamma(variable2, variable1) = out_gamma(variable1, variable2);
+            out_indicator(variable1, variable2) *= iteration;
+            out_indicator(variable1, variable2) += indicator(variable1, variable2);
+            out_indicator(variable1, variable2) /= iteration + 1;
+            out_indicator(variable2, variable1) = out_indicator(variable1, variable2);
           }
 
           out_interactions(variable1, variable2) *= iteration;
@@ -1131,7 +1212,7 @@ List gibbs_sampler(IntegerMatrix observations,
 
 
   if(edge_selection == true) {
-    return List::create(Named("gamma") = out_gamma,
+    return List::create(Named("indicator") = out_indicator,
                         Named("interactions") = out_interactions,
                         Named("thresholds") = out_thresholds);
   } else {